#30 chore: move ui related files to presentation module
When A depends on B which uses dataBinding, A also must configure dataBinding
ref: https://stackoverflow.com/questions/41765114/databinding-not-working-in-module
