+++ app/build.gradle
@@ -1,7 +1,6 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
-    id("androidx.navigation.safeargs.kotlin") // safeArgs
 }
 
 android {
@@ -48,11 +47,6 @@ dependencies {
 
     // kakao login
     implementation "com.kakao.sdk:v2-user:2.10.0"
-    // navigation
-    implementation("androidx.navigation:navigation-fragment-ktx:2.4.2")
-    implementation("androidx.navigation:navigation-ui-ktx:2.4.2")
-    // concatAdapter
-    implementation "androidx.recyclerview:recyclerview:1.3.0-alpha02"
 
     implementation project(":data")
     implementation project(":domain")
