+++ app/src/main/AndroidManifest.xml
@@ -3,8 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     package="com.kldaji.loancounselor">
 
-    <uses-permission android:name="android.permission.INTERNET" />
-
     <application
         android:name=".LoanCounselorApplication"
         android:allowBackup="true"
@@ -16,33 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.LoanCounselor"
         tools:targetApi="31">
-        <activity
-            android:name=".MainActivity"
-            android:exported="false" />
-        <activity
-            android:name=".LoginActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
-            android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW" />
-
-                <category android:name="android.intent.category.DEFAULT" />
-                <category android:name="android.intent.category.BROWSABLE" />
-
-                <!-- Redirect URI: "kakao${NATIVE_APP_KEY}://oauth" -->
-                <data
-                    android:host="oauth"
-                    android:scheme="@string/KAKAO_LOGIN_REDIRECT_URL" />
-            </intent-filter>
-        </activity>
     </application>
 
 </manifest>
