--- app/src/main/java/com/kldaji/loancounselor/Client.kt
@@ -1,3 +1,0 @@
-package com.kldaji.loancounselor
-
-class Client
