+++ app/src/main/java/com/kldaji/loancounselor/LoanCounselorApplication.kt
@@ -2,6 +2,7 @@ package com.kldaji.loancounselor
 
 import android.app.Application
 import com.kakao.sdk.common.KakaoSdk
+import com.kldaji.presentation.R
 
 class LoanCounselorApplication: Application() {
     override fun onCreate() {
