+++ domain/src/main/java/com/kldaji/domain/Client.kt
@@ -1,4 +1,3 @@
 package com.kldaji.domain
 
-class MyClass {
-}
+class Client
