+++ presentation/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
+    id("androidx.navigation.safeargs.kotlin") // safeArgs
 }
 
 android {
@@ -27,6 +28,10 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+    buildFeatures {
+        dataBinding true
+        viewBinding true
+    }
 }
 
 dependencies {
@@ -37,6 +42,14 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+
+    // kakao login
+    implementation "com.kakao.sdk:v2-user:2.10.0"
+    // navigation
+    implementation("androidx.navigation:navigation-fragment-ktx:2.4.2")
+    implementation("androidx.navigation:navigation-ui-ktx:2.4.2")
+    // concatAdapter
+    implementation "androidx.recyclerview:recyclerview:1.3.0-alpha02"
 
     implementation project(":domain")
 }
