+++ presentation/src/main/AndroidManifest.xml
@@ -2,4 +2,35 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.kldaji.presentation">
 
+    <uses-permission android:name="android.permission.INTERNET" />
+
+    <application>
+        <activity
+            android:name="com.kldaji.presentation.ui.MainActivity"
+            android:exported="false" />
+        <activity
+            android:name="com.kldaji.presentation.ui.login.LoginActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity
+            android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
+
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
+
+                <!-- Redirect URI: "kakao${NATIVE_APP_KEY}://oauth" -->
+                <data
+                    android:host="oauth"
+                    android:scheme="@string/KAKAO_LOGIN_REDIRECT_URL" />
+            </intent-filter>
+        </activity>
+    </application>
 </manifest>
