+++ presentation/src/main/java/com/kldaji/presentation/ui/search/SearchFragment.kt
@@ -1,11 +1,11 @@
-package com.kldaji.loancounselor
+package com.kldaji.presentation.ui.search
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.kldaji.loancounselor.databinding.FragmentSearchBinding
+import com.kldaji.presentation.databinding.FragmentSearchBinding
 
 class SearchFragment : Fragment() {
     private var _binding: FragmentSearchBinding? = null
