+++ presentation/src/main/java/com/kldaji/presentation/ui/setting/SettingFragment.kt
@@ -1,22 +1,21 @@
-package com.kldaji.loancounselor
+package com.kldaji.presentation.ui.setting
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.kldaji.loancounselor.databinding.FragmentClientInfoBinding
+import com.kldaji.presentation.databinding.FragmentSettingBinding
 
-
-class ClientInfoFragment : Fragment() {
-    private var _binding: FragmentClientInfoBinding? = null
+class SettingFragment : Fragment() {
+    private var _binding: FragmentSettingBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
-        _binding = FragmentClientInfoBinding.inflate(inflater, container, false)
+        _binding = FragmentSettingBinding.inflate(inflater, container, false)
         return binding.root
     }
