+++ presentation/src/main/res/drawable/ic_action_24.xml
