+++ presentation/src/main/res/drawable/ic_add_client_24.xml
