+++ presentation/src/main/res/drawable/ic_meeting_24.xml
