+++ presentation/src/main/res/drawable/ic_navigate_24.xml
