+++ presentation/src/main/res/drawable/ic_setting_24.xml
