+++ presentation/src/main/res/drawable/kakao_login_large_narrow.png
