+++ presentation/src/main/res/drawable/kakao_login_medium_narrow.png
