+++ presentation/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginActivity">
+    tools:context="com.kldaji.presentation.ui.login.LoginActivity">
 
     <ImageButton
         android:id="@+id/btn_kakao_login"
