+++ presentation/src/main/res/layout/fragment_client_info.xml
@@ -3,11 +3,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SettingFragment">
+    tools:context="com.kldaji.presentation.ui.clientInfo.ClientInfoFragment">
 
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="Setting Fragment" />
+        android:text="ClientInfo Fragment" />
 
 </FrameLayout>
