+++ presentation/src/main/res/layout/fragment_clients.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ClientsFragment">
+    tools:context="com.kldaji.presentation.ui.clients.ClientsFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
