+++ presentation/src/main/res/layout/fragment_scheduled_clients.xml
@@ -3,11 +3,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ClientInfoFragment">
+    tools:context="com.kldaji.presentation.ui.scheduledClients.ScheduledClientsFragment">
 
     <TextView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="ClientInfo Fragment" />
+        android:layout_height="match_parent" />
 
 </FrameLayout>
