+++ presentation/src/main/res/layout/fragment_setting.xml
@@ -3,10 +3,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScheduledClientsFragment">
+    tools:context="com.kldaji.presentation.ui.setting.SettingFragment">
 
     <TextView
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent"
+        android:text="Setting Fragment" />
 
 </FrameLayout>
