+++ presentation/src/main/res/layout/item_scheduled_client.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="scheduledClientView"
-            type="com.kldaji.loancounselor.ScheduledClientView" />
+            type="com.kldaji.presentation.ui.clients.entity.ScheduledClientView" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
