+++ presentation/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/clientsFragment"
-        android:name="com.kldaji.loancounselor.ClientsFragment"
+        android:name="com.kldaji.presentation.ui.clients.ClientsFragment"
         android:label="fragment_clients"
         tools:layout="@layout/fragment_clients" >
         <action
@@ -25,22 +25,22 @@
     </fragment>
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.kldaji.loancounselor.SearchFragment"
+        android:name="com.kldaji.presentation.ui.search.SearchFragment"
         android:label="fragment_search"
         tools:layout="@layout/fragment_search" />
     <fragment
         android:id="@+id/clientInfoFragment"
-        android:name="com.kldaji.loancounselor.ClientInfoFragment"
+        android:name="com.kldaji.presentation.ui.clientInfo.ClientInfoFragment"
         android:label="fragment_client_info"
         tools:layout="@layout/fragment_client_info" />
     <fragment
         android:id="@+id/settingFragment"
-        android:name="com.kldaji.loancounselor.SettingFragment"
+        android:name="com.kldaji.presentation.ui.setting.SettingFragment"
         android:label="fragment_setting"
         tools:layout="@layout/fragment_setting" />
     <fragment
         android:id="@+id/scheduledClientsFragment"
-        android:name="com.kldaji.loancounselor.ScheduledClientsFragment"
+        android:name="com.kldaji.presentation.ui.scheduledClients.ScheduledClientsFragment"
         android:label="fragment_scheduled_clients"
         tools:layout="@layout/fragment_scheduled_clients" >
         <argument
